tests: py: Add test for ambiguity while setting the value
This test checks bug identified and fixed in the commit mentioned below
In a statement if there are  multiple src data then it  would be
totally ambiguous to decide which value to set.

Before the commit was made it returned 134(BUG), but now it returns 1
i.e, an error message.

Following rules tests ambiguity while setting the value:

 $ sudo nft add rule ip test-ip4 output ct mark set {0x11333, 0x11}
 <cmdline>:1:41-55: Error: you cannot use a set here, unknown value to use
 add rule ip test-ip4 output ct mark set {0x11333, 0x11}
                             ~~~~~~~~~~~~^^^^^^^^^^^^^^^

Test: 986dea8 ("evaluate: avoid reference to multiple src data in
statements which set values")
Signed-off-by: Shyam Saini <mayhs11saini@gmail.com>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>