tests: py: Fail test forcefully when bug is not fixed
When we have "fail" in the test cases then py test doesn't complain
anything, but the test should complain if the fix is not applied.

Before applying 986dea8a4a9d ("evaluate: avoid reference to multiple src
data in statements which set values"), nft throws following error
message and exits with error code 134.

$ nft add rule x y tcp dport set { 0 , 1 }
  BUG: unknown expression type set reference
  nft: netlink_linearize.c:696: netlink_gen_expr: Assertion `0' failed.
  Aborted

This commit enforces nft-test.py to throw error message when the fix
is not applied.

Signed-off-by: Shyam Saini <mayhs11saini@gmail.com>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>